+++ packages/intents-sdk/src/bridges/poa-bridge/poa-bridge-utils.ts
@@ -69,11 +69,12 @@ const caip2Mapping = {
 	[Chains.Starknet]: "starknet:mainnet",
 	[Chains.Aleo]: "aleo:mainnet",
 	[Chains.Dash]: "dash:mainnet",
+	[Chains.Plasma]: "eth:9745",
 } satisfies Record<
 	string,
 	(typeof poaBridge.PoaBridgeNetworkReference)[Exclude<
 		keyof typeof poaBridge.PoaBridgeNetworkReference,
-		"NEAR" | "POLYGON" | "BSC" | "MONAD"
+		"NEAR" | "POLYGON" | "BSC" | "MONAD" | "FOGO"
 	>]
 >;
 
@@ -105,6 +106,7 @@ const tokenPrefixMapping = {
 	starknet: Chains.Starknet,
 	aleo: Chains.Aleo,
 	dash: Chains.Dash,
+	plasma: Chains.Plasma,
 };
 
 export function contractIdToCaip2(contractId: string): Chain {